pyproject.toml
[project]
name = "aws-rds-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Ramos",email = "matheussoaresramos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
